add support for custom build directory in simulation environment

+++ moo_sim_interface/simulation_environment_apis/open_modelica_simulator.py
@@ -16,12 +16,11 @@ from moo_sim_interface.utils.yaml_config_parser import prepare_simulation_enviro
 
 def run_simulation(return_results: bool = False, **args) -> Union[None, list]:
     if find_spec('OMPython') is None:
-        # try to install the OMPython package
         simulator_path = args.get('simulator_path')
-        install_openmodelica_package(simulator_path)
+        install_openmodelica_package(simulator_path)  # try to install the OMPython package
 
     (model_filename, model_path, input_values, input_names, num_chunks, final_names, sync_execution,
-     time_modulo, result_transformation) = prepare_simulation_environment(args)
+     time_modulo, result_transformation, custom_build_dir) = prepare_simulation_environment(args)
 
     if version("OMPython") > "3.4.9":
         model_path = model_path.as_posix()
@@ -53,8 +52,8 @@ def run_simulation(return_results: bool = False, **args) -> Union[None, list]:
 
     if num_chunks == 1:
         from moo_sim_interface.utils.OMPythonFast import ModelicaSystemFast
-
-        model = ModelicaSystemFast(model_path, model_name, commandLineOptions='--demoMode')
+        model = ModelicaSystemFast(model_path, model_name, commandLineOptions='--demoMode',
+                                   customBuildDirectory=custom_build_dir)
         for script in pre_sim_scripts:
             res = model.getconn.execute("runScript(\"" + script + "\")")
             if "Failed" in res:
@@ -72,7 +71,7 @@ def run_simulation(return_results: bool = False, **args) -> Union[None, list]:
         combined_results = run_simulation_in_parallel(final_names, indices, input_names, input_values, method,
                                                       model_path, model_name, start_time, step_size, stop_time,
                                                       tolerance, num_chunks, result_transformation,
-                                                      pre_sim_scripts, post_sim_scripts)
+                                                      pre_sim_scripts, post_sim_scripts, custom_build_dir)
 
     processed_results = post_simulation_data_processor.do_post_processing(args, input_values, combined_results,
                                                                           model_name, return_results=return_results)
@@ -101,8 +100,8 @@ def run_simulation_in_order(final_names, indices, initial_names, input_values, m
 
 
 def run_simulation_in_parallel(final_names, indices, initial_names, input_values, method, model_path, model_name,
-                               start_time, step_size, stop_time, tolerance, num_chunks,
-                               result_transformation, pre_sim_scripts, post_sim_scripts) -> list[list]:
+                               start_time, step_size, stop_time, tolerance, num_chunks, result_transformation,
+                               pre_sim_scripts, post_sim_scripts, custom_build_dir) -> list[list]:
     print(f'Running simulation in parallel with {num_chunks} workers.')
 
     batch_size = ceil(len(indices) / num_chunks)  # calculate the batch size and work on all batches in parallel
@@ -112,7 +111,7 @@ def run_simulation_in_parallel(final_names, indices, initial_names, input_values
     # Prepare the arguments to be passed to each worker
     worker_args = [
         (indices, final_names, initial_names, input_values, method, model_path, model_name, start_time, step_size,
-         stop_time, tolerance, pre_sim_scripts, post_sim_scripts)
+         stop_time, tolerance, pre_sim_scripts, post_sim_scripts, custom_build_dir)
         for indices in batched_indices]
 
     with multiprocessing.Pool(processes=num_chunks) as pool:
@@ -129,8 +128,9 @@ def run_simulation_in_parallel(final_names, indices, initial_names, input_values
 
 
 def simulate_model_worker(indices, final_names, initial_names, input_values, method, model_path, model_name,
-                          start_time, step_size, stop_time, tolerance, pre_sim_scripts, post_sim_scripts):
-    model, build_dir = create_omc_process(indices, model_path, model_name, pre_sim_scripts)
+                          start_time, step_size, stop_time, tolerance, pre_sim_scripts, post_sim_scripts,
+                          custom_build_dir):
+    model, build_dir = create_omc_process(indices, model_path, model_name, pre_sim_scripts, custom_build_dir)
 
     collected_results = []
 
@@ -159,20 +159,20 @@ def simulate_model_worker(indices, final_names, initial_names, input_values, met
 
     stop_omc_process(model)
 
-    try:
-        shutil.rmtree(build_dir)  # Remove the build directory
-    except Exception as e:
-        print(f"Error removing build directory {build_dir}: {e}")
+    if custom_build_dir is None:
+        try:
+            shutil.rmtree(build_dir)  # Remove the build directory
+        except Exception as e:
+            print(f"Error removing build directory {build_dir}: {e}")
 
     return indices, collected_results
 
 
-def create_omc_process(indices, model_path, model_name, pre_sim_scripts):
+def create_omc_process(indices, model_path, model_name, pre_sim_scripts, custom_build_dir):
     from moo_sim_interface.utils.OMPythonFast import ModelicaSystemFast
 
-    # create one tmp build dir for multiple indices
-    build_dir = construct_build_dir(model_name, indices)
-    os.mkdir(build_dir)
+    # create a temporary build dir for multiple indices
+    build_dir = construct_build_dir(model_name, indices, custom_build_dir)
     model = ModelicaSystemFast(model_path, model_name, customBuildDirectory=build_dir, commandLineOptions='--demoMode')
     for script in pre_sim_scripts:
         res = model.getconn.execute("runScript(\"" + script + "\")")
@@ -200,6 +200,7 @@ def construct_resultfile_name(model_name, index):
     return f'{model_name}_{index}.mat'.replace(' ', '_')
 
 
-def construct_build_dir(model_name, indices):
+def construct_build_dir(model_name, indices, custom_build_dir=None):
     index_appendix = f'{indices[0]}_{indices[-1]}' if len(indices) > 1 else f'{indices[0]}'
-    return os.path.join(os.getcwd(), f'{model_name}_{index_appendix}'.replace(' ', '_'))
+    base_dir = custom_build_dir if custom_build_dir is not None else os.getcwd()
+    return os.path.join(base_dir, f'{model_name}_{index_appendix}'.replace(' ', '_'))
